fix: 러닝목표 설정시 단위를 통일시키기 계산(m) -> 표시(km), 계산(sec) -> 표시(min)

+++ TrackUs/TrackUs/Sources/Commons/Model/Course.swift
@@ -17,7 +17,7 @@ struct Course: Decodable, Hashable {
     let content: String
     let courseRoutes: [GeoPoint]
     let distance: Double
-    let estimatedTime: Int
+    let estimatedTime: Double
     let participants: Int
     let runningStyle: String
     let startDate: Date
